Use consistent syntax
The delete method didn't create a $sql varaible but in-lined the string straight into the query method. This was out of line with every other method & changed for consistency.
Also the id parameter was not properly specified and would have created an error. Finally there was a typo in the comment